chore: refactor getCurrentUser
